+++ cdn/main.tf
@@ -164,7 +164,7 @@ resource "aws_cloudfront_distribution" "standard" {
   }
 
   lifecycle {
-    # Use `ignore_changes` to allow rotation without Terraform overwriting the value
+    # Use `ignore_changes` to allow custom_header secret rotation without Terraform overwriting the value
     ignore_changes = [origin]
   }