+++ postgres/database-dump/variables.tf
@@ -9,3 +9,12 @@ variable "environment" {
 variable "database_name" {
   type = string
 }
+
+variable "tasks" {
+  type = list(object({
+    from         = string
+    to           = string
+    from_account = optional(string)
+    to_account   = optional(string)
+  }))
+}